requirements/local.txt
-r common.txt

django-extensions==1.7.5
django-debug-toolbar==1.8